added cargo selection page and separated content from home page
