+++ src/App.jsx
@@ -1,23 +1,28 @@
-import react from "react"
-import { useState } from 'react'
+import React from 'react';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import './App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import HomePage from './pages/HomePage';
-import './App.css'
+import Register from './components/Registration/Register';
+import CargoPage from './pages/CargoPage';
 
 function App() {
-  const [count, setCount] = useState(0)
-
   return (
     <Router>
       <div>
         <Routes>
           <Route path="/" element={<HomePage />} />
+          <Route path="/register" element={<Register />} />
+          <Route path="/cargo" element={<CargoPage />} />
         </Routes>
       </div>
     </Router>
   );
 }
 
-export default App
+const rootElement = document.getElementById('root');
+const root = ReactDOM.createRoot(rootElement);
+root.render(<App />);
+
+export default App;
